+++ UI/src/components/Parsers/Config2Form/extractInfoSiblings.js
@@ -5,7 +5,9 @@ import determineDiseaseStatus from "./determineDiseaseStatus";
 
 export default function extractInfoSiblings(paramsObject, config){
     // Retrieve Params
-    var sampleIDs=paramsObject.pedigreeMapping.siblings.filter(id => id !== "");
+    var sampleIDs=paramsObject.pedigreeMapping.siblings.filter(function(id){
+        return id !== "" && determinePositionSampleID(id,paramsObject["sample.ids"]) !== -1;
+    });
     var indicesOfID=sampleIDs.map(function(d){
         return determinePositionSampleID(d,paramsObject["sample.ids"]);
     });